feat: 유저 삭제시에 유저가 존재하는지 확인

+++ src/main/java/com/seohamin/fshs/v2/domain/user/service/UserService.java
@@ -134,6 +134,13 @@ public class UserService implements UserDetailsService {
      */
     @Transactional
     public void deleteUser(final Long userId) {
+
+        // 1) 유저 존재 확인
+        if (!userRepository.existsById(userId)) {
+            throw new CustomException(ExceptionCode.USER_NOT_EXIST);
+        }
+
+        // 2) 삭제
         userRepository.deleteById(userId);
     }
